Moved the packages out of pkg to simplify the structure a bit
