+++ affinityinjector/affinityinjector_test.go
